+++ quill/include/quill/detail/LogMacros.h
@@ -30,7 +30,7 @@
 // Main Log Macros
 
 #if defined(QUILL_NOFN_MACROS)
-// clang-format off
+  // clang-format off
 #define QUILL_LOGGER_CALL_NOFN(likelyhood, logger, log_statement_level, fmt, ...)                  \
   do {                                                                                             \
     struct {                                                                                       \
@@ -39,11 +39,12 @@
                                                 "n/a", fmt, log_statement_level}; }                \
       } anonymous_log_record_info;                                                                 \
                                                                                                    \
-    if (likelyhood(logger->should_log<log_statement_level>()))                                     \
+    if (likelyhood(logger->template should_log<log_statement_level>()))                            \
     {                                                                                              \
       constexpr bool try_fast_queue {true}; /* Available in dual queue mode only */                \
       constexpr bool is_backtrace_log_record {false};                                              \
-      logger->log<try_fast_queue, is_backtrace_log_record, decltype(anonymous_log_record_info)>    \
+      logger->template log<try_fast_queue, is_backtrace_log_record,                                \
+                           decltype(anonymous_log_record_info)>                                    \
                                                              (FMT_STRING(fmt),  ##__VA_ARGS__);    \
     }                                                                                              \
   } while (0)
@@ -60,11 +61,12 @@
                                                 function_name, fmt, log_statement_level}; }        \
       } anonymous_log_record_info;                                                                 \
                                                                                                    \
-    if (likelyhood(logger->should_log<log_statement_level>()))                                     \
+    if (likelyhood(logger->template should_log<log_statement_level>()))                            \
     {                                                                                              \
       constexpr bool try_fast_queue {true}; /* Available in dual queue mode only */                \
       constexpr bool is_backtrace_log_record {false};                                              \
-      logger->log<try_fast_queue, is_backtrace_log_record, decltype(anonymous_log_record_info)>    \
+      logger->template log<try_fast_queue, is_backtrace_log_record,                                \
+                           decltype(anonymous_log_record_info)>                                    \
                                                              (FMT_STRING(fmt),  ##__VA_ARGS__);    \
     }                                                                                              \
   } while (0)
@@ -78,11 +80,12 @@
                                                 function_name, fmt, quill::LogLevel::Backtrace}; } \
       } anonymous_log_record_info;                                                                 \
                                                                                                    \
-    if (QUILL_LIKELY(logger->should_log<quill::LogLevel::Backtrace>()))                            \
+    if (QUILL_LIKELY(logger->template should_log<quill::LogLevel::Backtrace>()))                   \
     {                                                                                              \
       constexpr bool try_fast_queue {true}; /* Available in dual queue mode only */                \
       constexpr bool is_backtrace_log_record {true};                                               \
-      logger->log<try_fast_queue, is_backtrace_log_record, decltype(anonymous_log_record_info)>    \
+      logger->template log<try_fast_queue, is_backtrace_log_record,                                \
+                           decltype(anonymous_log_record_info)>                                    \
                                                              (FMT_STRING(fmt),  ##__VA_ARGS__);    \
     }                                                                                              \
   } while (0)