+++ quill/include/quill/handlers/FileHandler.h
@@ -14,6 +14,7 @@ namespace quill
 
 enum class FilenameAppend
 {
+  DateTime,
   Date,
   None
 };